EnergyModelsGeography

EnergyModelsGeography — Module

Main module for EnergyModelsGeography.jl.

Exports:

Area, RefArea, GeoAvailability, Transmission, RefStatic, RefDynamic, PipeMode, PipeSimple, PipeLinepackSimple.

The Julia package extends EnergyModelsBase with functionality to set up an energy system consisting of several separate regions, with transmissions between.

The extension focuses on overriding the function EMB.create_model to add additional types, variables, and constraints to the model.

Manual outline

Quick Start

Philosophy

Optimization variables

Constraint functions

TransmissionModes

Adding investments

Examples

How to guides

Contribute to EnergyModelsGeography

Library outline

Public interface

Reference