EnergymodelsInvestments extensions

EnergyModelsGeography requires significantly less direct interaction with EnergyModelsInvestments compared to EnergyModelsBase. In practice, we only provide a legacy constructor. All other functionality is handled within the internal functions.

Investment data

InvestmentData types

Transmission mode investments utilize the same investment data type (SingleInvData) as investments in node capacities.

Legacy constructors

We provide a legacy constructor, TransInvData, that uses the same input as in version 0.5.x. If you want to adjust your model to the latest changes, please refer to the section Update your model to the latest version of EnergyModelsInvestments.

EnergyModelsGeography.TransInvData — Function

TransInvData(;
    capex_trans::TimeProfile,
    trans_max_inst::TimeProfile,
    trans_max_add::TimeProfile,
    trans_min_add::TimeProfile,
    inv_mode::Investment = ContinuousInvestment(),
    trans_start::Union{Real, Nothing} = nothing,
    trans_increment::TimeProfile = FixedProfile(0),
    capex_trans_offset::TimeProfile = FixedProfile(0),
)

Legacy constructor for a TransInvData.

The new storage descriptions allows now for a reduction in functions which is used to make EnergModelsInvestments less dependent on EnergyModelsBase.

The core changes to the existing structure is the move of the required parameters to the type Investment (e.g., the minimum and maximum added capacity is only required for investment mdodes that require these parameters) as well as moving the lifetime to the type [LifetimeMode], when required.

See the documentation for further information regarding how you can translate your existing model to the new model.

TransInvData(;
    capex_trans::TimeProfile,
    trans_max_inst::TimeProfile,
    trans_max_add::TimeProfile,
    trans_min_add::TimeProfile,
    inv_mode::Investment = ContinuousInvestment(),
    trans_start::Union{Real, Nothing} = nothing,
    trans_increment::TimeProfile = FixedProfile(0),
    capex_trans_offset::TimeProfile = FixedProfile(0),
)

Legacy constructor for a InvData.

The new storage descriptions allows now for a reduction in functions which is used to make EnergModelsInvestments less dependent on EnergyModelsBase.

The core changes to the existing structure is the move of the required parameters to the type Investment (e.g., the minimum and maximum added capacity is only required for investment mdodes that require these parameters) as well as moving the lifetime to the type [LifetimeMode], when required.

See the documentation for further information regarding how you can translate your existing model to the new model.