Transmission

Transmission occurs on specified transmission corridors from one area to another. On each corridor, there can exist several TransmissionModes that are transporting resources using a range of technologies.

It is important to consider the from and to Area when specifying a Transmission corridor. The chosen direction has an influence on whether the variables $\texttt{trans\_in}[m, t]$ and $\texttt{trans\_out}[m, t]$ are positive or negative for exports in the case of bidirectional transport. This is also explained on the page Optimization variables.

Transmission types

EnergyModelsGeography.Transmission — Type

A Transmission corridor.

A geographic corridor where TransmissionModes are used to transport resources.

Fields

from::Area is the area resources are transported from.

to::Area is the area resources are transported to.

modes::Vector{<:Transmission} are the transmission modes that are available.

Functions for accessing fields of Transmission types

The following functions are defined for accessing fields from a Transmission as well as finding a subset of Transmission corridors:

EnergyModelsGeography.modes — Function

modes(l::Transmission)
modes(ℒᵗʳᵃⁿˢ::Vector{<:Transmission})

Return an array of the transmission modes for a transmission corridor l or for a vector of transmission corridors ℒᵗʳᵃⁿˢ.

EnergyModelsGeography.modes_sub — Function

modes_sub(ℳ::Vector{<:TransmissionMode}, string::String)

Returns all transmission modes that include in the name the string.

modes_sub(ℳ::Vector{<:TransmissionMode}, string_array::Array{String})

Returns all transmission modes that include in the name all entries of the array string_array.

modes_sub(l::Transmission, mode_type::TransmissionMode)
modes_sub(ℒᵗʳᵃⁿˢ::Vector{<:Transmission}, mode_type::TransmissionMode)

Return an array containing all TransmissionModes of type mode_type in Transmission corridor l or for a vector of transmission corridors ℒᵗʳᵃⁿˢ.

modes_sub(l::Transmission, p::Resource)

Return an array containing all TransmissionModes that transport the resource p in Transmission corridor l or in a vector of transmission corridors ℒᵗʳᵃⁿˢ.

EnergyModelsGeography.corr_from — Function

corr_from(from::Area, ℒᵗʳᵃⁿˢ)
corr_from(from::String, ℒᵗʳᵃⁿˢ)

Returns all transmission corridors in ℒᵗʳᵃⁿˢ that orginate in the Area from. If from is provided as String, it returns the corridors in which the name is equal to from

EnergyModelsGeography.corr_to — Function

corr_to(to::Area, ℒᵗʳᵃⁿˢ)
corr_to(to::String, ℒᵗʳᵃⁿˢ)

Returns all transmission corridors in ℒᵗʳᵃⁿˢ that end in the Area to. If to is provided as String, it returns the corridors in which the name is equal to to

EnergyModelsGeography.corr_from_to — Function

corr_from_to(from::Union{Area,String}, to::Union{Area,String}, ℒᵗʳᵃⁿˢ)

Returns the transmission corridor that orginate in the Area from and end in the Area to.

The function accepts both inputs as String and Area as well as a combination of both

EnergyModelsGeography.modes_of_dir — Function

modes_of_dir(ℒ, dir::Int)

Return the transmission modes of dir directions for transmission modes ℳ.

modes_of_dir(l, dir::Int)

Return the transmission modes of dir directions for transmission corridor l`.