EnergyModelsGeography.jl

EnergyModelsGeography — Module

Main module for EnergyModelsGeography.jl.

Exports:

Area, RefArea, GeoAvailability, Transmission, RefStatic, RefDynamic, PipeMode, PipeSimple, PipeLinepackSimple.

The Julia package extends EnergyModelsBase.jl with functionality to set up an energy system consisting of several separate regions, with transmissions between.

The extension focuses on overriding the function EMB.create_model to add additional types, variables, and constraints to the model.

Manual outline

Quick Start

Philosophy

General design philosophy

Extensions to the model

Optimization variables

Area

TransmissionMode

Constraint functions

Capacity constraints

Transmission loss functions

Balance functions

Operational expenditure constraints

TransmissionModes

Examples

The package is installed with ] add

The code was downloaded with git clone

Library outline

Public interface

Area

Transmission

TransmissionMode

Reference

Index

Types

Methods