Transmission

Transmission occurs on specified transmission corridors from one area to another. On each corridor, there can exist several TransmissionModes that are transporting resources using a range of technologies.

It is important to consider the from and to Area when specifying a Transmission corridor. The chosen direction has an influence on whether the variables $\texttt{trans\_in}[m, t]$ and $\texttt{trans\_out}[m, t]$ are positive or negative for exports in the case of bidirectional transport. This is also explained on the page Optimization variables.

Transmission types

EnergyModelsGeography.Transmission — Type

Transmission <: AbstractElement

A geographic corridor between two Areas where TransmissionModes are used to transport resources.

Fields

from::Area is the area resources are transported from.

to::Area is the area resources are transported to.

modes::Vector{<:Transmission} are the transmission modes that are available.

Functions for accessing fields of Transmission types

The following functions are defined for accessing fields from a Transmission as well as finding a subset of Transmission corridors:

EnergyModelsGeography.modes — Function

modes(l::Transmission)
modes(ℒᵗʳᵃⁿˢ::Vector{<:Transmission})

Returns an array of the transmission modes for a transmission corridor l or for a vector of transmission corridors ℒᵗʳᵃⁿˢ.

EnergyModelsGeography.modes_sub — Function

modes_sub(ℳ::Vector{<:TransmissionMode}, str::String)
modes_sub(ℳ::Vector{<:TransmissionMode}, str_arr::Array{String})

Returns an array containing all TransmissionModes that include in the name the String str or any of values in the String array str_arr.

modes_sub(l::Transmission, mode_type::TransmissionMode)
modes_sub(ℒᵗʳᵃⁿˢ::Vector{<:Transmission}, mode_type::TransmissionMode)

Returns an array containing all TransmissionModes of type mode_type in [Transmission])@ref) corridor l or for a vector of transmission corridors ℒᵗʳᵃⁿˢ.

modes_sub(l::Transmission, p::Resource)
modes_sub(ℒᵗʳᵃⁿˢ::Vector{<:Transmission}, p::Resource)

Returns an array containing all TransmissionModes that transport the resource p in [Transmission])@ref) corridor l or in a vector of transmission corridors ℒᵗʳᵃⁿˢ.

modes_sub(l::Transmission, str::String)
modes_sub(l::Transmission, string_array::Array{String})
modes_sub(ℒᵗʳᵃⁿˢ::Vector{<:Transmission}, str::String)
modes_sub(ℒᵗʳᵃⁿˢ::Vector{<:Transmission}, string_array::Array{String})

Returns an array containing all TransmissionModes of [Transmission])@ref) corridor l or in a vector of transmission corridors ℒᵗʳᵃⁿˢ that include in the name the String str or any of values in the String array str_arr.

EnergyModelsGeography.corr_from — Function

corr_from(from::Area, ℒᵗʳᵃⁿˢ)
corr_from(from::String, ℒᵗʳᵃⁿˢ)

Returns all transmission corridors in ℒᵗʳᵃⁿˢ that orginate in the Area from. If from is provided as String, it returns the corridors in which the name is equal to from

EnergyModelsGeography.corr_to — Function

corr_to(to::Area, ℒᵗʳᵃⁿˢ)
corr_to(to::String, ℒᵗʳᵃⁿˢ)

Returns all transmission corridors in ℒᵗʳᵃⁿˢ that end in the Area to. If to is provided as String, it returns the corridors in which the name is equal to to

EnergyModelsGeography.corr_from_to — Function

corr_from_to(from::Union{Area,String}, to::Union{Area,String}, ℒᵗʳᵃⁿˢ)

Returns the transmission corridor that orginate in the Area from and end in the Area to.

The function accepts both inputs as String and Area as well as a combination of both